backend/requirements.txt
djangorestframework  # for rest api
djangorestframework-simplejwt # for jwt authentication
pyyaml # for yaml parsing
requests # for making http
django-cors-headers # for cors
channels # for websockets
channels-redis # for redis
django_celery_results # for celery results
django_celery_beat # for celery beat
celery # for celery
django>=4.2.0 # for django
django-channels # for channels
selenium # for selenium
python-dotenv # for env
pyotp # for otp
africastalking # for africastalking
fcm-django  # for push notifications
Pillow # for image processing
asyncio # for async
boto3 # for aws
django-storages # for aws
python-magic # for file validation